cellmap to kt, squash with MutableFieldState
